requirements.txt
pyyaml
pytest>=2.6.0